Started to move spectrum code to single location. Removed O0 from lcd code, please test, works very, very well with SPI. Added files to all build systems.
