+++ mchf-eclipse/drivers/ui/lcd/ui_lcd_hy28.h
@@ -116,14 +116,16 @@ void 	UiLcdHy28_SendByteSpi(uint8_t byte);
 uint8_t UiLcdHy28_ReadByteSpi(void);
 void 	UiLcdHy28_WriteDataOnly( unsigned short data);
 
+void 	UiLcdHy28_BulkPixel_Put(uint16_t pixel);
 void 	UiLcdHy28_BulkWrite(uint16_t* pixels, uint32_t len);
 void 	UiLcdHy28_BulkWriteColor(uint16_t color, uint32_t len);
 void 	UiLcdHy28_CloseBulkWrite(void);
 
 void	UiLcdHy28_WriteRAM_Prepare(void);
 void	UiLcdHy28_WriteRAM_Finish(void);
+void 	UiLcdHy28_SetCursorA( unsigned short Xpos, unsigned short Ypos );
 
-void 	UiLcdHy28_DrawSpectrum(q15_t *fft,ushort color,ushort shift);
+void 	UiLcdHy28_DrawSpectrum_Prepare(q15_t *fft_old, q15_t *fft_new, const ushort color_old, const ushort color_new, const ushort shift);
 void 	UiLcdHy28_DrawSpectrum_Interleaved(q15_t *fft_old, q15_t *fft_new, const ushort color_old, const ushort color_new, const ushort shift);
 
 void 	UiLcdHy28_Test(void);
