+++ mchf-eclipse/drivers/ui/lcd/ui_spectrum.h
@@ -1,0 +1,17 @@
+#ifndef __UI_SPECTRUM_H
+#define __UI_SPECTRUM_H
+/************************************************************************************
+**                                                                                 **
+**                               mcHF QRP Transceiver                              **
+**                             K Atanassov - M0NKA 2014                            **
+**                                                                                 **
+**---------------------------------------------------------------------------------**
+**                                                                                 **
+**  File name:                                                                     **
+**  Description:                                                                   **
+**  Last Modified:                                                                 **
+**  Licence:      CC BY-NC-SA 3.0                                                **
+************************************************************************************/
+
+
+#endif
