+++ mchf-eclipse/drivers/ui/ui_driver.c
@@ -6023,6 +6023,7 @@ static void UiDriverInitSpectrumDisplay(void)
 
 	// Ready
 	sd.enabled		= 1;
+	sd.first_run 	= 2;
 }
 
 //
